api/requirements.txt
Django==4.2.0
django-ninja==1.1.0
django-cors-headers==4.3.1
pydantic==2.4.0
python-dotenv==1.0.0
aiofiles==23.1.0
httpx==0.24.1
waitress==3.0.2
